Actions: Escape quoted content in paths (#3026)
When quoting is enabled, the quotes in paths were previously not
escaped. This led to improper handling for paths with quotes and/or
backslashes.

For double quoted content, escape backslashes and double quotes with
backslashes.

For single quoted content, replace single quotes with a close quote, an
escaped single quote, and a starting single quote.