App chooser: Quote custom executable file paths
The MIME application chooser only escapes spaces currently when picking
a custom executable file. However, the picked file paths may contain
non-space special characters which also need to be escaped to be
interpreted correctly with sh rules, like backslashes. Make use of the
g_shell_quote function to quote the path instead, which is guaranteed to
escape file paths for the shell correctly.