nemo-window-menus.c: Recalculate extension/action items at an earlier stage in opening the file manu.
The 'activated' signal is too late to adjust the visible entries
in the File popup menu - it can end up being the wrong size for
the items finally shown.

Using the 'show' signal of the menu itself to hide/show conditional
items - this occurs just before the widget size will be calculated.

ref:
https://github.com/linuxmint/mint21.1-beta/issues/34