nemo-view-dnd.c: Fix the check for a web link's title, and use the url if it's missing.
Firefox can send a url terminated by a newline (\n), tricking nemo
into thinking there's also a title, but it's really just an empty
string. Check for that as well as NULL.

Use the url as a substitute for the title, instead of the file's
basename, as this stays consistent with behavior in other browsers
(and firefox in other situations).

ref: #3173