nemo-application.c: Update the custom css provider when the system theme changes.
Nemo needs some special style info for certain UI elements, so at
startup we check the current gtk theme for that support. If there
is none, the system theme's aggregate CSS is loaded and injected
with additional classes, then loaded as a style provider to replace
it.

When the user changes themes, this work needs to be reversed, and
the theme evaluated for nemo support again.