Revert "nemo-desktop: use app_id as .desktop file name (#3340)"
This reverts commit 4aa7a7b725fb9934c4c7187431b7200d11253c45.

The easiest way to find the desktop file for Nemo is to call it
nemo.desktop.

Not any of the following:

org.Nemo
com.Nemo
io.Nemo
com.cinnamon.Nemo
com.linuxmint.cinnamon.Nemo

This creates regressions since it breaks any ref to the app (namely
people adding it as a pin in GWL) and it serves no purpose nor fixes
any bugs we're aware of.