eel-gnome-extensions.c: Support whitespace in execution commands (#3413)
A default command line with whitespace will fail to execute, either
silently or with a visible error, depending on the trigger. In addition,
some actions would default to launching with GNOME Terminal even when
another terminal was configured as the default. This adds a default list
of terminals, proper switches for execution, and a boolean for whether
or not the command line sequence needs to be escaped.

Closes #2571