nemo-action-manager.c: Generate unique names for actions when they're loaded.
The action name is internal, and used by the UI manager only.
When actions are added, they're presented in the order. Updates
to the UI manager only update action existence at a given level,
to prevent needless widget turnover.

The problem occurs when the action layout is modified: If an
action is moved only up or down at its current depth, the UI
manager won't change the order visibly.

By making each generation of actions unique from the previous,
this makes sure the actions are always newly added in the correct
order.