nemo-window.c: Remember the second pane's location until the remaining pane's location changes.
If the user is working in the split view, and toggles split pane
off (from either side), remember the location of the closed
pane and restore it if the split view is toggled again.

If the remaining pane's location changes, the stored location is
cleared, and traditional behavior will prevail (opening the second
pane in the same location as the first).

If a pane had multiple tabs open, only the active tab's location
is remembered and re-opened.

ref:
https://blog.linuxmint.com/?p=4850#comment-499214