nemo-search-engine-advanced.c: Support wildcard mimetypes.
This allows a search helper to process multiple types of the same
family, such as 'image/*'.

Note that for now, a more specific search helper will be used
over a wildcard helper, regardless of priority.  If a helper
handles 'image/jpeg', it will be used exclusively, *not* the
'image/*' helper.