nemo-action-manager.c: Don't randomize action names when rebuilding.
GtkAction names are 'interned' and never freed for the life of the
process. For Nemo actions, the name (which is really an escaped url)
is made 'unique' every time actions are reloaded, in order to avoid
issues with the UI manager when editing the layout.

When a lot of actions are in use, this can end up using a lot of
memory over time as each instance of a GtkAction adds to Gtk's
internal name list.

Since this is all the fault of using a custom action layout,
use part of that layout file's mtime to stamp each generation of
actions. If this never changes for the life of the process (the
user doesn't edit the layout), the internal action names will
never change, and just be re-used.