Revert "add --select compatibility flag (#3589)"
This reverts commit e21b2846642cf8664739665817c165d25db65acb.

It makes no sense to add a flag which does nothing just so scripts
can call it.

If Caja, Thunar, Dolphin and Nautilus committed to some standard,
interface or common specification, we'd commit as well but in the
meantime we're not going to have empty features or flags in here just
because they do.

This is adding unnecessary complexity into Nemo just to save the trouble
of some if() statement in some external script. It's not justified.