nemo-places-sidebar.c: Restore visibility of mounts with no associated volume.
We started skipping these in 93815617c16, but there are some
legitimate instances where these should be shown, and are still
shown in other areas of Nemo (the desktop, the move/copy-to
submenus).

One case is Virtualbox shared folders. If sharing a host folder
with a guest and no mount point is provided, guest-additions
mounts this in /media and it shows up in the GVolumeMonitor (and
desktop, etc...). If a mount-point *is* provided, and that folder
exists on the guest, the mount is *not* visible in a GVolumeMonitor,
and will not be visible in Nemo, except as a normal folder. This
seems logical.

Another seems to be cifs/samba mounts, depending on how they're
mounted. Using smb:// uris causes these to appear in the Network
section, but using mount they no longer appear in the sidebar
in any section.

The other part of that commit, skipping shadowed mounts, remains
correct - and these other places mentioned above were already
filtering these out as well.

Ref:
linuxmint/mint22.3-beta#67.